requirements.txt
python_dotenv
langchain
openai
chainlit
chromadb
tiktoken
lxml
llama_index
html2text